feat: US-P0h - Replace Win32 Time APIs with std::chrono
Replace GetLocalTime/SYSTEMTIME/GetTickCount64/FileTimeToLocalFileTime/
SystemTimeToFileTime in domain code with std::chrono and ctime equivalents.

- meos_util: add getLocalTm/mkgmtime helpers; change all SYSTEMTIME API
  signatures to std::tm; replace FILETIME arithmetic with chrono math
- TimeStamp: replace FILETIME/SystemTimeToFileTime with mkgmtime+gmtime_s
  preserving exact FILETIME-epoch arithmetic for DB compatibility
- oEvent/oEventSpeaker: GetLocalTime → getLocalTm(); timing → steady_clock
- iof30interface: SystemTimeToTzSpecificLocalTime → mkgmtime+localtime_s
- onlineinput: simplify using mkgmtime directly from std::tm
- autotask/SportIdent/oEventSQL/MeosSQL: GetTickCount → steady_clock helpers
- methodeditor: srand(GetTickCount) → chrono-based seed
- onlineresults/mysqldaemon: GetTickCount64 → steady_clock

+++ code/oEvent.h
@@ -966,7 +966,7 @@ protected:
   bool calculateTeamResults(vector<const oTeam*> &teams, int leg, ResultType resultType);
   void calculateModuleTeamResults(const set<int> &cls, vector<oTeam *> &teams);
 
-  unsigned int lastTimeConsistencyCheck = 0;
+  uint64_t lastTimeConsistencyCheck = 0;
   mutable bool lastResultCalcPrelState = false;
   mutable bool lastResultCalcSplitResult = false;
 public:
